// src/app/profile/dtos/load-person-dto.ts
export class LoadPersonDTO {
  id: string;
  middleName: string;
  suffix: string;
  age: number;
  ethnicity: string;
  religiousPreference: string;
  department: string;
  command: string;
  gtcTrainingDate: Date;
  adamsTrainingDate: Date;
  hasCompletedAWARE: boolean;
  supervisor: string;
  workCenter: string;
  workRoom: string;
  shift: string;
  jobTitle: string;
  eaos: Date;
  prd: Date;
  dateOfDeparture: Date;
  billetAssignment: string;
  paygrade: string;
  designation: string;
  uic: string;
  dutyStatus: string;
  sex: string;
  lastName: string;
  firstName: string;
  ssn: string;
  doDId: string;
  dateOfBirth: Date;
  dateOfArrival: Date;
  division: string;
}